+++ src/payload-config/i18n/locales/en.json
@@ -260,7 +260,7 @@
     "checkIn": "Check in",
     "checkingIn": "Checking in...",
     "alreadyCheckedIn": "Already checked in",
-    "checkInSuccessful": "Check-in Successful!",
+    "showTicket": "Show this screen to receive ticket!",
     "confirmedSuccessfully": "Confirmed successfully",
     "zone": "Zone:",
     "ticketCode": "Ticket Code:",
