+++ src/payload-config/i18n/locales/vi.json
@@ -260,7 +260,7 @@
     "checkIn": "Check-in",
     "checkingIn": "Đang check-in...",
     "alreadyCheckedIn": "Đã check-in truớc đó",
-    "checkInSuccessful": "Check-in Thành công!",
+    "showTicket": "Hãy đưa màn hình cho nhân viên để nhận vé!",
     "confirmedSuccessfully": "Đã xác nhận thành công!",
     "zone": "Khu vực:",
     "ticketCode": "Mã vé:",
